refactor[api-server]: 컨트롤러테스트 공통 필드를 갖는 추상 클래스로 변경
