--- src/test/java/com/watermelon/server/annotations/ControllerTest.java
@@ -1,24 +1,0 @@
-package com.watermelon.server.annotations;
-
-import com.watermelon.server.MockLoginInterceptorConfig;
-import com.watermelon.server.randomevent.controller.LotteryController;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Import;
-import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.TYPE)
-@Retention(RetentionPolicy.RUNTIME)
-@AutoConfigureMockMvc
-@AutoConfigureRestDocs
-@MockBean(JpaMetamodelMappingContext.class)
-@Import(MockLoginInterceptorConfig.class)
-public @interface ControllerTest {
-}
