+++ src/test/java/com/watermelon/server/randomevent/link/controller/LinkControllerTest.java
@@ -1,41 +1,30 @@
 package com.watermelon.server.randomevent.link.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.watermelon.server.Constants;
+import com.watermelon.server.ControllerTest;
 import com.watermelon.server.DocumentConstants;
-import com.watermelon.server.MockLoginInterceptorConfig;
-import com.watermelon.server.annotations.ControllerTest;
-import com.watermelon.server.randomevent.controller.LotteryController;
+import com.watermelon.server.annotations.ControllerTestAnno;
 import com.watermelon.server.randomevent.link.dto.MyLinkDto;
 import com.watermelon.server.randomevent.link.service.LinkService;
 import com.watermelon.server.randomevent.link.utils.LinkUtils;
-import com.watermelon.server.randomevent.service.LotteryService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.context.annotation.Import;
-import org.springframework.http.HttpMethod;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static com.epages.restdocs.apispec.MockMvcRestDocumentationWrapper.document;
 import static com.watermelon.server.Constants.*;
 import static com.watermelon.server.common.constants.PathConstants.MY_LINK;
 import static com.watermelon.server.common.constants.PathConstants.SHORTED_LINK;
-import static org.junit.jupiter.api.Assertions.*;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.request;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-import org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders;
-
-@ControllerTest
 @WebMvcTest(LinkController.class)
-class LinkControllerTest {
+class LinkControllerTest extends ControllerTest {
 
     //TODO 추상클래스 상속 구조로 변경
     @Autowired
