+++ src/test/java/com/watermelon/server/randomevent/parts/controller/PartsControllerTest.java
@@ -1,8 +1,7 @@
 package com.watermelon.server.randomevent.parts.controller;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
+import com.watermelon.server.ControllerTest;
 import com.watermelon.server.DocumentConstants;
-import com.watermelon.server.annotations.ControllerTest;
 import com.watermelon.server.randomevent.parts.dto.response.ResponseMyPartsListDto;
 import com.watermelon.server.randomevent.parts.dto.response.ResponsePartsDrawDto;
 import com.watermelon.server.randomevent.parts.dto.response.ResponseRemainChanceDto;
@@ -11,10 +10,8 @@ import com.watermelon.server.randomevent.parts.service.PartsService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.web.servlet.MockMvc;
 
 import static com.watermelon.server.Constants.*;
 import static com.watermelon.server.Constants.TEST_TOKEN;
@@ -24,23 +21,15 @@ import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.docu
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.patch;
 import java.util.List;
 
-import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-@ControllerTest
 @WebMvcTest(PartsController.class)
-class PartsControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
+class PartsControllerTest extends ControllerTest {
 
     @MockBean
     private PartsService partsService;
-
-    @Autowired
-    private ObjectMapper objectMapper;
 
     @Test
     @DisplayName("파츠 뽑기 결과를 반환한다.")
